+++ packages/gallery/src/components/item/imageItem.js
@@ -1,3 +1,4 @@
+/* eslint-disable prettier/prettier */
 import React from 'react';
 import { GALLERY_CONSTS, utils, isSEOMode } from 'pro-gallery-lib';
 import { GalleryComponent } from '../galleryComponent';
@@ -156,10 +157,7 @@ export default class ImageItem extends GalleryComponent {
           }
         : {};
 
-      if (
-        !isHighResImageLoaded &&
-        (gotFirstScrollEvent || settings.forceImagePreload)
-      ) {
+      if (!isHighResImageLoaded && gotFirstScrollEvent) {
         let preload = null;
         const preloadProps = {
           className: 'gallery-item-visible gallery-item gallery-item-preloaded',
@@ -259,3 +257,4 @@ export default class ImageItem extends GalleryComponent {
     return renderedItem;
   }
 }
+/* eslint-enable prettier/prettier */