+++ packages/lib/src/common/defaultStyles.js
@@ -115,13 +115,13 @@ const defaultStyles = {
   itemShadowSize: 10,
   imageLoadingMode: GALLERY_CONSTS.loadingMode.BLUR,
   // imageLoadingColor: startWithColor: 'color-3'
-  expandAnimation: GALLERY_CONSTS.expandAnimations.NO_EFFECT,
   // oneColorAnimationColor: startWithColor: 'color-1',
   videoSound: false,
   videoSpeed: '1',
   videoLoop: true,
   jsonStyleParams: '',
   showVideoControls: false,
+  shouldIndexDirectShareLinkInSEO: true,
 };
 
 /* Object.entries(galleryOptions).forEach(([styleParam, settings]) => {