+++ packages/lib/src/settings/options/index.js
@@ -87,6 +87,7 @@ import allowLeanGallery from './allowLeanGallery';
 import placeGroupsLtr from './placeGroupsLtr';
 import pauseAutoSlideshowOnHover from './pauseAutoSlideshowOnHover';
 import showVideoControls from './showVideoControls';
+import shouldIndexDirectShareLinkInSEO from './shouldIndexDirectShareLinkInSEO';
 
 export default {
   itemClick,
@@ -178,11 +179,11 @@ export default {
   placeGroupsLtr,
   showVideoControls,
   pauseAutoSlideshowOnHover,
+  shouldIndexDirectShareLinkInSEO,
 };
 
 // TODO = add the options:
 /*
-allowLeanGallery
 itemOpacity
 arrowsColor
 imageLoadingColor